SSL session resumption, add memcached timeout option and max blocked worker process option.
The memcached_set and memcached_get operation during ssl handshake will be blocked at most 5 seconds by default,
with this option, we can specify a smaller time like 1000ms, set it to 0 or don't set it to disable this option.
The memcached_set and memcached_get operation may block all the nginx worker processes, then the nginx can't do anything,
with this option, we can specify maximum worker process can be concurrently blocked, the left will go through handshake directly.
set it to 0 or don't set it to disable this option.

 ssl_session_cache    memcached:ssl:localhost:11211:1000:2;